alx_travel_app/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
django-cors-headers
drf-yasg